+++ src/main/java/com/widyu/fcm/api/FcmController.java
@@ -1,12 +1,10 @@
 package com.widyu.fcm.api;
 
 import com.widyu.fcm.api.dto.FcmSendDto;
+import com.widyu.fcm.api.dto.response.FcmNotificationResponses;
 import com.widyu.fcm.api.dto.response.FcmSendResponse;
 import com.widyu.fcm.application.FcmService;
 import com.widyu.global.response.ApiResponseTemplate;
-import com.widyu.global.util.MemberUtil;
-import com.widyu.member.domain.Member;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +15,7 @@ import java.io.IOException;
 @RestController
 @RequestMapping("/api/v1/fcm")
 @RequiredArgsConstructor
-public class FcmController implements FcmDocs{
+public class FcmController implements FcmDocs {
 
     private final FcmService fcmService;
 
@@ -29,5 +27,13 @@ public class FcmController implements FcmDocs{
                 .code("FCM_2001")
                 .message("푸시 메시지 전송 성공")
                 .body(response);
+    }
+
+    @GetMapping()
+    public ApiResponseTemplate<FcmNotificationResponses> getNotification() {
+        return ApiResponseTemplate.ok()
+                .code("FCM_2002")
+                .message("사용자 알림 조회 성공")
+                .body(fcmService.getNotificationsForCurrentUser());
     }
 }
