+++ src/main/java/com/widyu/fcm/domain/repository/FcmNotificationRepository.java
@@ -1,8 +1,10 @@
 package com.widyu.fcm.domain.repository;
 
 import com.widyu.fcm.domain.FcmNotification;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FcmNotificationRepository extends JpaRepository<FcmNotification, Long> {
+    List<FcmNotification> findAllByMemberFcmToken_MemberIdOrderByCreatedAtDesc(Long id);
 }
